+++ tomcat/src/main/java/com/techcourse/controller/HomeController.java
@@ -1,9 +1,9 @@
 package com.techcourse.controller;
 
 import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http.HttpRequest;
-import org.apache.coyote.http.HttpResponse;
-import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.response.StatusCode;
 
 public class HomeController extends AbstractController {
