+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -7,9 +7,9 @@ import java.util.Objects;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http.HttpRequest;
-import org.apache.coyote.http.HttpResponse;
-import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.response.StatusCode;
 
 public class RegisterController extends AbstractController {
