+++ tomcat/src/main/java/org/apache/catalina/controller/AbstractController.java
@@ -1,8 +1,8 @@
 package org.apache.catalina.controller;
 
-import org.apache.coyote.http.HttpRequest;
-import org.apache.coyote.http.HttpResponse;
-import org.apache.coyote.http.RequestMethod;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.request.RequestMethod;
 
 public abstract class AbstractController implements Controller {
