+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.controller;
 
-import org.apache.coyote.http.HttpRequest;
-import org.apache.coyote.http.HttpResponse;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
 
 public interface Controller {
