+++ tomcat/src/main/java/org/apache/catalina/controller/NotFoundController.java
@@ -1,9 +1,8 @@
 package org.apache.catalina.controller;
 
-import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http.HttpRequest;
-import org.apache.coyote.http.HttpResponse;
-import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.response.StatusCode;
 
 public class NotFoundController extends AbstractController {
