+++ tomcat/src/main/java/org/apache/coyote/http/response/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.response;
 
 public enum StatusCode {
